# src/public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Mini Note RESTful API",
        "license": {
            "name": "Apache"
        },
        "version": "0.1.0"
    },
    "servers": [
        {
            "url": "http://localhost:2880/api/v1",
            "description": "API server development"
        },
        {
            "url": "https://mininote.orlangur.link/api/v1",
            "description": "API server production"
        }
    ],
    "paths": {
...
        "/notes": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Notes list",
                "description": "View notes information",
                "operationId": "notes.list",
                "parameters": [
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Note offset",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Note limit",
                        "required": false,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Notes"
                ],
                "summary": "Create note",
                "description": "Create note",
                "operationId": "notes.create",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "note"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title note",
                                        "type": "string"
                                    },
                                    "note": {
                                        "description": "Note",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status note",
                                        "type": "string",
                                        "default": "draft",
                                        "enum": [
                                            "draft",
                                            "public",
                                            "archive"
                                        ]
                                    },
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/notes/{id}": {
            "get": {
                "tags": [
                    "Notes"
                ],
                "summary": "Read note",
                "description": "Read note",
                "operationId": "notes.read",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Notes"
                ],
                "summary": "Update note",
                "description": "Update note",
                "operationId": "notes.update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "title",
                                    "note"
                                ],
                                "properties": {
                                    "title": {
                                        "description": "Title note",
                                        "type": "string"
                                    },
                                    "note": {
                                        "description": "Note",
                                        "type": "string"
                                    },
                                    "status": {
                                        "description": "Status note",
                                        "type": "string",
                                        "default": "draft",
                                        "enum": [
                                            "draft",
                                            "public",
                                            "archive"
                                        ]
                                    },
                                    "category_id": {
                                        "description": "Category ID",
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Notes"
                ],
                "summary": "Delete note",
                "description": "Delete note",
                "operationId": "notes.delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Note id",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/UniversalDTO"
                                    }
                                }
                            }
                        }
                    },
                    "default": {
                        "description": "unexpected error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "UniversalDTO": {
                "title": "UniversalDTO",
                "required": [],
                "properties": {
                    "success": {
                        "type": "boolean"
                    },
                    "status": {
                        "type": "integer"
                    },
                    "error": {
                        "properties": {
                            "code": {
                                "type": "integer",
                                "example": "10"
                            },
                            "message": {
                                "type": "string",
                                "example": "Error message"
                            }
                        },
                        "type": "object"
                    },
                    "data": {
                        "type": "array",
                        "items": {}
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}